+++ src/refactoring/features/cart/components/CartList.tsx
@@ -1,6 +1,5 @@
 import { CartItem, Coupon } from '../../../../types';
-import { useDiscountCalculator } from '../../../hooks/useDiscountCalculator';
-import { formatDiscountRate, formatCurrency } from '../../../utils';
+import { useDiscountCalculator } from '../hooks/useDiscountCalculator';
 import { CartCard } from './CartCard';
 import { CartCoupon } from './CartCoupon';
 import { CartTotal } from './CartTotal';
@@ -26,18 +25,6 @@ export const CartList = ({
 
   const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } = cartCalculations;
 
-  const getAppliedDiscount = (item: CartItem) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-      if (quantity >= discount.quantity) {
-        appliedDiscount = Math.max(appliedDiscount, discount.rate);
-      }
-    }
-    return appliedDiscount;
-  };
-
   return (
     <div>
       <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
@@ -46,7 +33,6 @@ export const CartList = ({
           <CartCard
             key={item.product.id}
             item={item}
-            getAppliedDiscount={getAppliedDiscount}
             updateQuantity={updateQuantity}
             removeFromCart={removeFromCart}
           />
